+++ tests/unit/test_app.py
@@ -111,6 +111,10 @@ def test_can_call_to_dict_on_current_request(sample_app):
     # out a few keys as a basic sanity test.
     assert response['method'] == 'GET'
     assert response['json_body'] == {}
+    # We also want to verify that to_dict() is always
+    # JSON serializable so we check we can roundtrip
+    # the data to/from JSON.
+    assert isinstance(json.loads(json.dumps(response)), dict)
 
 
 def test_will_pass_captured_params_to_view(sample_app):
@@ -233,6 +237,22 @@ def test_json_body_available_on_multiple_content_types():
     assert raw_body == '{"foo": "bar"}'
 
 
+def test_json_body_available_with_lowercase_content_type_key():
+    demo = app.Chalice('demo-app')
+
+    @demo.route('/', methods=['POST'])
+    def index():
+        return (demo.current_request.json_body, demo.current_request.raw_body)
+
+    event = create_event_with_body({'foo': 'bar'})
+    del event['params']['header']['Content-Type']
+    event['params']['header']['content-type'] = 'application/json'
+
+    json_body, raw_body = demo(event, context=None)
+    assert json_body == {'foo': 'bar'}
+    assert raw_body == '{"foo": "bar"}'
+
+
 def test_content_types_must_be_lists():
     demo = app.Chalice('app-name')
 
@@ -333,3 +353,12 @@ def test_chalice_view_errors_propagate_in_debug_mode(sample_app):
     event = create_event('/notfound', 'GET', {})
     with pytest.raises(NotFoundError):
         sample_app(event, context=None)
+
+
+def test_case_insensitive_mapping():
+    mapping = app.CaseInsensitiveMapping({'HEADER': 'Value'})
+
+    assert mapping['hEAdEr']
+    assert mapping.get('hEAdEr')
+    assert 'hEAdEr' in mapping
+    assert repr({'header': 'Value'}) in repr(mapping)